move test class LongIntScanningFactory to separate file
